+++ src/Connection/Driver/MariaDBCapabilities.php
@@ -63,8 +63,8 @@ class MariaDBCapabilities implements DriverCapabilities
     /**
      * @inheritDoc
      */
-    public function getIdentifierQuote(): string
+    public function hasRenameColumn(): bool
     {
-        return '`';
+        return version_compare($this->driverInfo->getVersion(), '10.5.2', '>=');
     }
 }